Ruff is toying with me

+++ simpunch/level1.py
@@ -366,7 +366,7 @@ def add_starfield_clear(input_data: NDCube) -> NDCube:
 def generate_dummy_polarization(map_scale: float = 0.225,
                                 pol_factor: float = 0.5) -> NDCube:
     """Create a synthetic polarization map."""
-    shape = [int(floor(180 / map_scale)), int(floor(360 / map_scale))]
+    shape = [floor(180 / map_scale), floor(360 / map_scale)]
     xcoord = np.linspace(-pol_factor, pol_factor, shape[1])
     ycoord = np.linspace(-pol_factor, pol_factor, shape[0])
     xin, yin = np.meshgrid(xcoord, ycoord)
